only:int rpm -> float rpm

+++ firmware/controllers/algo/accel_enrichment.cpp
@@ -34,7 +34,7 @@ float TpsAccelEnrichment::getTpsEnrichment() {
 		// If disabled, return 0.
 		return 0;
 	}
-	int rpm = Sensor::getOrZero(SensorType::Rpm);
+	float rpm = Sensor::getOrZero(SensorType::Rpm);
 	if (rpm == 0) {
 		return 0;
 	}
